requirements.txt
pandas
numpy
requests
torch
transformers
yfinance
scikit-learn
matplotlib
python-dotenv